hotfix

+++ Golden_Neighbour/scripts/php/payments/payments.php
@@ -32,13 +32,18 @@ if (mail($to, $subject, $message, $headers)) {
         die("Connection failed: " . mysqli_connect_error());
     }
     $cart_id = $_POST['cart_id'];
+    $sql = "DELETE FROM cart WHERE selected_seat = $selected_seat AND ";
+    $result = mysqli_query($conn, $sql);
+
+    if ($result === FALSE) {
+        die("Error executing query: " . $conn->error);
+    }
     $sql = "DELETE FROM cart WHERE cart_id = $cart_id";
     $result = mysqli_query($conn, $sql);
 
     if ($result === FALSE) {
         die("Error executing query: " . $conn->error);
     }
-    $conn = mysqli_connect("localhost", "root", "", "goldenneighbour");
 
 
     mysqli_close($conn);
